删除Settings.增加N5的英文文本和暂定遗物图片
